backend/foodgram/requirements.txt
--upgrade pip
django==2.2.19
psycopg2-binary==2.8.6 
python-dotenv
pytz==2020.1
sqlparse==0.3.1 
requests==2.26.0
Pillow
djangorestframework==3.12.4
django-cors-headers
djoser
djangorestframework-simplejwt==4.7.2
django-filter